requirements.txt
pandas
requests
httpx
databricks-sql-connector